Extract component AppErrorPane
It renders the error with the possible formats that come from the backend.
